- hide premium services from default filter at >8000ms delay - fixed GUI deadlock by synchronizing pack() and updateValues() - plausibility check before showing best bounds (needed if not all InfoServices send best bounds)

+++ src/jap/TrustModel.java
@@ -555,7 +555,7 @@ public class TrustModel extends BasicTrustModel implements IXMLEncodable
 		model = new TrustModel(MSG_SERVICES_WITH_COSTS, 2);
 		model.setAttribute(PaymentAttribute.class, TRUST_IF_TRUE);
 		model.setAttribute(ExpiredCertsAttribute.class, TRUST_RESERVED);
-		model.setAttribute(DelayAttribute.class, TRUST_IF_AT_MOST, new Integer(4000));
+		model.setAttribute(DelayAttribute.class, TRUST_IF_AT_MOST, new Integer(8000));
 		model.setAttribute(SpeedAttribute.class, TRUST_IF_AT_LEAST, new Integer(100));
 		model.setAttribute(NumberOfMixesAttribute.class, TRUST_IF_AT_LEAST, new Integer(3));
 		TRUST_MODEL_PREMIUM = model;
